+++ questevent-backend/src/main/java/com/questevent/service/ProgramWalletTransactionService.java
@@ -2,11 +2,10 @@ package com.questevent.service;
 
 import com.questevent.entity.Program;
 import com.questevent.entity.User;
-import jakarta.transaction.Transactional;
 
 public interface ProgramWalletTransactionService {
 
-    void creditGems(User user, Program program, int amount);
+    void creditGems(User user, Program program, Long amount);
 
     void autoSettleExpiredProgramWallets();
